Version 1.5.0
PLEASE DELETE YOUR CONFIG/MESSAGES.YML! AND DO A BACKUP OF THE SHOPS AND MESSAGES SO CHANGES CAN APPLY!

Changelogs:
- Removed /delshop command
- Removed /setshop command
- Removed /shopvisit command

Added in the config the max shops limit by default and the permission to limit the max shops of a player, this is how it works:
If you put in the config the limit of 10, the players without the permission and OP (SHOULD!) have a shop limit of 10 then players with the permission playershop.limit.<Number between 0 and 10(or the max default in the config), should have a lower max shops limit, for example.
You write in the config a limit of 10, then you add to the default "USER" role the permission playershop.limit.5, they can make only 5 shops.

Now there're only 3 commands,
- /shops, this command open a GUI, with 2 buttons, one to see your own shops and add them, another to see how many shops you have and how many you can make.
- /shoplogo <Block_ID> <Shop_Number>, this command let you change the block "logo" of your shops.
- /playershop, this command let you see the commands.

Permissions:

- /shops, no permissions needed.
- Permission to use the setshop button: playershop.setshop
- Permission to use the /playershop command: playershop.setshop
- Permission to limit the playershops: playershop.limit.<Number>